MM-48640 Move migrateRecentEmojis out of mattermost-redux
